Update voxel.py
In this updated script, the --maxlength argument is added to specify the maximum recording length in minutes. This value is converted to seconds and stored in the pdat.max_recording_length attribute. The StreamProcessor class checks if the maximum recording length has been reached and stops and starts a new recording if necessary.